Update home.html
Added link to Js File as well as jquery script tag. 
Large overlays working when hovered over center of image with text in the middle as well. 
Need to solve what color the intro text will be (If i'm going to even have any). 
Will need to add footer and maybe section under pictures. 
Need to Add links in the overlays to respective sections.